fix(payments-next): Page - revise Checkout layout
