+++ apps/payments/next/app/[locale]/[offeringId]/checkout/layout.tsx
@@ -2,21 +2,72 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+import { headers } from 'next/headers';
+
+import { PurchaseDetails, TermsAndPrivacy } from '@fxa/payments/ui/server';
+import { getLocaleFromRequest } from '@fxa/shared/l10n';
+
+import { getCartData, getContentfulContent } from '../../../_lib/apiClient';
+
 // TODO - Replace these placeholders as part of FXA-8227
 export const metadata = {
   title: 'Mozilla accounts',
   description: 'Mozilla accounts',
 };
 
+interface CheckoutParams {
+  cartId: string;
+  locale: string;
+  interval: string;
+  offeringId: string;
+}
+
 export interface CheckoutSearchParams {
   experiment?: string;
   promotion_code?: string;
 }
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
+  params,
 }: {
   children: React.ReactNode;
+  params: CheckoutParams;
 }) {
-  return <>{children}</>;
+  const headersList = headers();
+  const locale = getLocaleFromRequest(
+    params,
+    headersList.get('accept-language')
+  );
+  const contentfulData = getContentfulContent(params.offeringId, locale);
+  const cartData = getCartData(params.cartId);
+  const [contentful, cart] = await Promise.all([contentfulData, cartData]);
+
+  return (
+    <>
+      <header className="page-title-container">
+        <h1 className="page-header">Under Construction</h1>
+      </header>
+
+      <section className="payment-panel" aria-label="Purchase details">
+        <PurchaseDetails
+          interval={cart.interval}
+          locale={locale}
+          invoice={cart.nextInvoice}
+          purchaseDetails={contentful.purchaseDetails}
+        />
+      </section>
+
+      <div className="page-body rounded-t-lg tablet:rounded-t-none">
+        {children}
+        <TermsAndPrivacy
+          locale={locale}
+          {...cart}
+          {...contentful.commonContent}
+          {...contentful.purchaseDetails}
+          showFXALinks={true}
+        />
+      </div>
+    </>
+  );
 }
